Issue #SC-2212 feat: Fixed the classcast from Double to Int for pkgVersion

+++ publish-pipeline/content-publish/src/main/scala/org/sunbird/job/content/publish/domain/Event.scala
@@ -22,8 +22,8 @@ class Event(eventMap: java.util.Map[String, Any], partition: Int, offset: Long) 
   def objectType: String = readOrDefault[String]("edata.metadata.objectType", "")
 
   def pkgVersion: Double = {
-    val pkgVersion = readOrDefault[Int]("edata.metadata.pkgVersion", 0)
-    pkgVersion.toDouble
+    val pkgVersion = readOrDefault[Double]("edata.metadata.pkgVersion", 0)
+    pkgVersion
   }
 
   def validEvent(config: ContentPublishConfig): Boolean = {
